Remove TestInternalDLR
