+++ .travis.yml
@@ -24,9 +24,6 @@ script:
   - dotnet restore Dlr.sln --packages ./packages
   - make release
   - pushd bin/$CONFIGURATION
-  - mono ./net40/TestInternalDLR.exe
-  - mono ./net45/TestInternalDLR.exe
-  - dotnet ./netcoreapp2.0/TestInternalDLR.dll
   - mono ./net40/Metadata.exe
   - mono ./net45/Metadata.exe
   - dotnet ./netcoreapp2.0/Metadata.dll
