+++ Dlr.sln
@@ -1,13 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26711.1
+VisualStudioVersion = 15.0.27130.2010
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{9C454D45-326F-4EA1-8177-64341267A17F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RowanTest.Common", "Tests\RowanTest.Common\RowanTest.Common.csproj", "{7B67CCAD-3C88-4A1E-8F5C-6AD784DFB52F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestInternalDLR", "Tests\TestInternalDLR\TestInternalDLR.csproj", "{75B776E7-2246-4E02-98A9-A99D3DA4DB54}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClrAssembly", "Tests\ClrAssembly\ClrAssembly.csproj", "{BEE737B9-18D5-48D9-8672-9A896213C98B}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -52,14 +48,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B67CCAD-3C88-4A1E-8F5C-6AD784DFB52F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B67CCAD-3C88-4A1E-8F5C-6AD784DFB52F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B67CCAD-3C88-4A1E-8F5C-6AD784DFB52F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B67CCAD-3C88-4A1E-8F5C-6AD784DFB52F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{75B776E7-2246-4E02-98A9-A99D3DA4DB54}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75B776E7-2246-4E02-98A9-A99D3DA4DB54}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{75B776E7-2246-4E02-98A9-A99D3DA4DB54}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{75B776E7-2246-4E02-98A9-A99D3DA4DB54}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BEE737B9-18D5-48D9-8672-9A896213C98B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BEE737B9-18D5-48D9-8672-9A896213C98B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BEE737B9-18D5-48D9-8672-9A896213C98B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -93,8 +81,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{7B67CCAD-3C88-4A1E-8F5C-6AD784DFB52F} = {9C454D45-326F-4EA1-8177-64341267A17F}
-		{75B776E7-2246-4E02-98A9-A99D3DA4DB54} = {9C454D45-326F-4EA1-8177-64341267A17F}
 		{BEE737B9-18D5-48D9-8672-9A896213C98B} = {9C454D45-326F-4EA1-8177-64341267A17F}
 		{848BBEB0-63E6-4736-B60C-23A9D733593D} = {9C454D45-326F-4EA1-8177-64341267A17F}
 		{60056F49-17D6-4C5C-80EC-F203BA2F7E3B} = {BFEE2666-74E5-49B0-BEB7-1F01CECD7F98}
