--- Tests/TestInternalDLR/Main.cs
@@ -1,28 +1,0 @@
-/* ****************************************************************************
- *
- * Copyright (c) Microsoft Corporation. All rights reserved.
- *
- * This source code is subject to terms and conditions of the Apache License, Version 2.0. A 
- * copy of the license can be found in the License.html file at the root of this distribution. If 
- * you cannot locate the  Apache License, Version 2.0, please send an email to 
- * dlr@microsoft.com. By using this source code in any fashion, you are agreeing to be bound 
- * by the terms of the Apache License, Version 2.0.
- *
- * You must not remove this notice, or any other, from this software.
- *
- *
- * ***************************************************************************/
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-using RowanTest.Common;
-
-namespace TestInternalDLR {
-    public class MainRunner {
-        static void Main() {
-            //delegate all work to RowanTest.Common.Runner
-            Runner.RunAssembly(System.Reflection.Assembly.GetExecutingAssembly());
-        }
-    }
-}
