--- Tests/TestInternalDLR/TestInternalDLR.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="$(ProjectDir)..\..\Build\Common.proj" />
-  <PropertyGroup>
-    <TargetFrameworks Condition="'$(TargetFrameworks)' == ''">net40;net45;netcoreapp2.0</TargetFrameworks>
-    <PlatformTarget>anycpu</PlatformTarget>
-    <ProjectGuid>{75B776E7-2246-4E02-98A9-A99D3DA4DB54}</ProjectGuid>
-    <OutputType>Exe</OutputType>
-    <RootNamespace>TestInternalDLR</RootNamespace>
-    <AssemblyName>TestInternalDLR</AssemblyName>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Src\Microsoft.Dynamic\Microsoft.Dynamic.csproj" />
-    <ProjectReference Include="..\RowanTest.Common\RowanTest.Common.csproj" />
-  </ItemGroup>
-</Project>
